requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==3.2
idna==2.10
jedi==0.17.2
packaging==20.4
parso==0.7.1
pip==21.0.1
pyasn1==0.4.8
pycparser==2.20
pyOpenSSL==19.1.0
pyparsing==2.4.6
PySocks==1.7.1
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
